# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"NagVijay Smart Monitoring System","description":"\n        ## NagVijay Smart Monitoring System (NSMS)\n        \n        Enterprise Monitoring Platform for India Post.\n        \n        **Features:**\n        - Google Login & JWT Authentication\n        - Role Based Access Control\n        - Office & Employee Master\n        - Target Engine with Google Forms/Sheets Integration\n        - Analytics & Reports\n        - AI Chatbot (Gemini)\n        - Notifications & Scheduler\n        - Audit Logs\n        \n        **Target:** Nagpur City Division (Scalable to Region, Circle, National)\n        \n        ### Authentication\n        Use `/api/v1/auth/google` for Google OAuth.\n        Include JWT in header: `Authorization: Bearer <token>`\n        \n        ### Roles\n        - SUPER_ADMIN: Full system access\n        - DIVISION_ADMIN: Division-level management\n        - OFFICE_ADMIN: Office-level management\n        - EMPLOYEE: Own data view/edit\n        - AUDITOR: Read-only + audit logs\n        ","contact":{"name":"NagVijay NSMS Team","email":"support@nagvijay.india-post"},"license":{"name":"MIT"},"version":"1.0.0-MVP"},"paths":{"/api/v1/health":{"get":{"tags":["Health"],"summary":"Health Check","description":"Comprehensive health check - DB, Supabase, Gemini, n8n","operationId":"health_check_api_v1_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/health/ready":{"get":{"tags":["Health"],"summary":"Readiness Probe","description":"Kubernetes readiness probe - lightweight","operationId":"readiness_check_api_v1_health_ready_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/health/live":{"get":{"tags":["Health"],"summary":"Liveness Probe","description":"Kubernetes liveness probe","operationId":"liveness_check_api_v1_health_live_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/auth/register":{"post":{"tags":["Authentication"],"summary":"Register with Email","operationId":"register_api_v1_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Register Api V1 Auth Register Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/login":{"post":{"tags":["Authentication"],"summary":"Login with Email","operationId":"login_api_v1_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Login Api V1 Auth Login Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/refresh":{"post":{"tags":["Authentication"],"summary":"Refresh Access Token","description":"Refresh access token using refresh token","operationId":"refresh_token_api_v1_auth_refresh_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshTokenRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Refresh Token Api V1 Auth Refresh Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/me":{"get":{"tags":["Authentication"],"summary":"Get Current User","description":"Get current authenticated user profile","operationId":"get_me_api_v1_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Get Me Api V1 Auth Me Get"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/auth/logout":{"post":{"tags":["Authentication"],"summary":"Logout","description":"Logout - client should discard tokens. Server could blacklist in future with Redis.","operationId":"logout_api_v1_auth_logout_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/offices":{"get":{"tags":["Offices"],"summary":"List Offices","description":"List offices with pagination, search, and filters","operationId":"list_offices_api_v1_offices_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"default":1,"title":"Page"}},{"name":"page_size","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":20,"title":"Page Size"}},{"name":"search","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Search"}},{"name":"office_type","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Office Type"}},{"name":"division","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Division"}},{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status"}},{"name":"sort_by","in":"query","required":false,"schema":{"type":"string","default":"created_at","title":"Sort By"}},{"name":"sort_order","in":"query","required":false,"schema":{"type":"string","pattern":"^(asc|desc)$","default":"desc","title":"Sort Order"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response List Offices Api V1 Offices Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["Offices"],"summary":"Create Office","operationId":"create_office_api_v1_offices_post","security":[{"HTTPBearer":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OfficeCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Create Office Api V1 Offices Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/offices/stats":{"get":{"tags":["Offices"],"summary":"Office Statistics by Type","operationId":"get_office_stats_api_v1_offices_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/offices/{office_id}":{"get":{"tags":["Offices"],"summary":"Get Office by ID","operationId":"get_office_api_v1_offices__office_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"office_id","in":"path","required":true,"schema":{"type":"string","title":"Office Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Get Office Api V1 Offices  Office Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Offices"],"summary":"Update Office","operationId":"update_office_api_v1_offices__office_id__put","security":[{"HTTPBearer":[]}],"parameters":[{"name":"office_id","in":"path","required":true,"schema":{"type":"string","title":"Office Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OfficeUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Update Office Api V1 Offices  Office Id  Put"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Offices"],"summary":"Delete Office","operationId":"delete_office_api_v1_offices__office_id__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"office_id","in":"path","required":true,"schema":{"type":"string","title":"Office Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Delete Office Api V1 Offices  Office Id  Delete"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/offices/bulk-import":{"post":{"tags":["Offices"],"summary":"Bulk Import Offices","operationId":"bulk_import_offices_api_v1_offices_bulk_import_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OfficeBulkImport"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Bulk Import Offices Api V1 Offices Bulk Import Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/daily-reports/":{"get":{"tags":["Daily Reports","Daily Reports"],"summary":"Get Daily Reports","operationId":"get_daily_reports_api_v1_daily_reports__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"report_date","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"description":"Filter by date (YYYY-MM-DD)","title":"Report Date"},"description":"Filter by date (YYYY-MM-DD)"},{"name":"office_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Office Id"}},{"name":"division","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"default":"Nagpur City","title":"Division"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DailyReportResponse"},"title":"Response Get Daily Reports Api V1 Daily Reports  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["Daily Reports","Daily Reports"],"summary":"Create Daily Report","operationId":"create_daily_report_api_v1_daily_reports__post","security":[{"HTTPBearer":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DailyReportCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DailyReportResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/daily-reports/summary":{"get":{"tags":["Daily Reports","Daily Reports"],"summary":"Get Daily Summary","operationId":"get_daily_summary_api_v1_daily_reports_summary_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"report_date","in":"query","required":true,"schema":{"type":"string","format":"date","description":"Date in YYYY-MM-DD","title":"Report Date"},"description":"Date in YYYY-MM-DD"},{"name":"division","in":"query","required":false,"schema":{"type":"string","default":"Nagpur City","title":"Division"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DailyReportSummary"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/daily-reports/export":{"get":{"tags":["Daily Reports","Daily Reports"],"summary":"Export Daily Report","operationId":"export_daily_report_api_v1_daily_reports_export_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"report_date","in":"query","required":true,"schema":{"type":"string","format":"date","description":"Date in YYYY-MM-DD","title":"Report Date"},"description":"Date in YYYY-MM-DD"},{"name":"format","in":"query","required":false,"schema":{"type":"string","pattern":"^(csv|json)$","default":"csv","title":"Format"}},{"name":"office_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Office Id"}},{"name":"division","in":"query","required":false,"schema":{"type":"string","default":"Nagpur City","title":"Division"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/daily-reports/non-reporting":{"get":{"tags":["Daily Reports","Daily Reports"],"summary":"Non-Reporting Offices for a Date","operationId":"get_non_reporting_offices_api_v1_daily_reports_non_reporting_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"report_date","in":"query","required":true,"schema":{"type":"string","format":"date","description":"Date in YYYY-MM-DD","title":"Report Date"},"description":"Date in YYYY-MM-DD"},{"name":"division","in":"query","required":false,"schema":{"type":"string","default":"Nagpur City","title":"Division"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/daily-reports/{report_id}":{"get":{"tags":["Daily Reports","Daily Reports"],"summary":"Get Daily Report By Id","operationId":"get_daily_report_by_id_api_v1_daily_reports__report_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"report_id","in":"path","required":true,"schema":{"type":"integer","title":"Report Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DailyReportResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/reports/generate":{"post":{"tags":["Reports"],"summary":"Generate Report","operationId":"generate_report_api_v1_reports_generate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReportRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Generate Report Api V1 Reports Generate Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/reports/dpr":{"get":{"tags":["Reports"],"summary":"Daily Performance Report","operationId":"get_dpr_api_v1_reports_dpr_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"division","in":"query","required":false,"schema":{"type":"string","default":"Nagpur City","title":"Division"}},{"name":"report_date","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"YYYY-MM-DD","title":"Report Date"},"description":"YYYY-MM-DD"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Get Dpr Api V1 Reports Dpr Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/reports/monthly":{"get":{"tags":["Reports"],"summary":"Monthly Consolidated Report","operationId":"get_monthly_report_api_v1_reports_monthly_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"financial_year","in":"query","required":true,"schema":{"type":"string","pattern":"^\\d{4}-\\d{2}$","title":"Financial Year"},"example":"2024-25"},{"name":"month","in":"query","required":true,"schema":{"type":"integer","maximum":12,"minimum":1,"title":"Month"}},{"name":"division","in":"query","required":false,"schema":{"type":"string","default":"Nagpur City","title":"Division"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Get Monthly Report Api V1 Reports Monthly Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/reports/export":{"get":{"tags":["Reports"],"summary":"Export Report","operationId":"export_report_api_v1_reports_export_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"report_type","in":"query","required":true,"schema":{"type":"string","pattern":"^(DAILY|MONTHLY|OFFICE_WISE|SCHEME_WISE)$","title":"Report Type"}},{"name":"format","in":"query","required":false,"schema":{"type":"string","pattern":"^(JSON|PDF|EXCEL|CSV)$","default":"JSON","title":"Format"}},{"name":"financial_year","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Financial Year"}},{"name":"division","in":"query","required":false,"schema":{"type":"string","default":"Nagpur City","title":"Division"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Export Report Api V1 Reports Export Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/integrations/forms/webhook":{"post":{"tags":["Integrations"],"summary":"Google Forms Webhook Receiver","description":"Receive webhook from Google Forms Apps Script.\nConverts form response to DailyOfficeReport.\nLogs errors to sync_errors table.","operationId":"forms_webhook_api_v1_integrations_forms_webhook_post","parameters":[{"name":"X-Webhook-Secret","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Webhook-Secret"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/integrations/sheets/status":{"get":{"tags":["Integrations"],"summary":"Google Sheets Status","operationId":"sheets_status_api_v1_integrations_sheets_status_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/integrations/sheets/read":{"post":{"tags":["Integrations"],"summary":"Read Google Sheet","description":"payload: { \"spreadsheet_id\": \"...\", \"range\": \"Sheet1!A1:Z100\" }","operationId":"read_sheet_api_v1_integrations_sheets_read_post","requestBody":{"content":{"application/json":{"schema":{"type":"object","title":"Payload"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/integrations/n8n/trigger":{"post":{"tags":["Integrations"],"summary":"Trigger n8n Workflow","description":"Manually trigger n8n workflow\npayload: { \"event\": \"notification\", \"webhook_path\": \"my-workflow\", \"data\": {...} }","operationId":"trigger_workflow_api_v1_integrations_n8n_trigger_post","requestBody":{"content":{"application/json":{"schema":{"type":"object","title":"Payload"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/integrations/sheets/sync-daily-office-report":{"post":{"tags":["Integrations"],"summary":"Sync Daily Office Report","operationId":"sync_daily_office_reports_endpoint_api_v1_integrations_sheets_sync_daily_office_report_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/integrations/status":{"get":{"tags":["Integrations"],"summary":"All Integrations Status","operationId":"all_integrations_status_api_v1_integrations_status_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/":{"get":{"tags":["Root"],"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"tags":["Root"],"summary":"Simple Health","operationId":"simple_health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"AnalyticsFilter":{"properties":{"financial_year":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Financial Year"},"division":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Division"},"region":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Region"},"office_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Office Id"},"employee_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Employee Id"},"scheme_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Scheme Id"},"scheme_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Scheme Type"},"start_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Start Date"},"end_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"End Date"},"period_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Period Type"},"office_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Office Type"}},"type":"object","title":"AnalyticsFilter"},"DailyReportCreate":{"properties":{"office_id":{"type":"string","title":"Office Id"},"office_name":{"type":"string","title":"Office Name"},"office_code":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Office Code"},"report_date":{"type":"string","format":"date","title":"Report Date"},"sb_opened":{"type":"integer","title":"Sb Opened","default":0},"sb_closed":{"type":"integer","title":"Sb Closed","default":0},"net_accounts":{"type":"integer","title":"Net Accounts","default":0},"pli_policies":{"type":"integer","title":"Pli Policies","default":0},"sum_assured":{"type":"number","title":"Sum Assured","default":0.0},"premium":{"type":"number","title":"Premium","default":0.0},"speed_post_document":{"type":"integer","title":"Speed Post Document","default":0},"speed_post_parcel":{"type":"integer","title":"Speed Post Parcel","default":0},"business_post":{"type":"integer","title":"Business Post","default":0},"logistics":{"type":"integer","title":"Logistics","default":0},"international_letter":{"type":"integer","title":"International Letter","default":0},"aadhaar_transactions":{"type":"integer","title":"Aadhaar Transactions","default":0},"aadhaar_amount":{"type":"number","title":"Aadhaar Amount","default":0.0}},"type":"object","required":["office_id","office_name","report_date"],"title":"DailyReportCreate"},"DailyReportResponse":{"properties":{"office_id":{"type":"string","title":"Office Id"},"office_name":{"type":"string","title":"Office Name"},"office_code":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Office Code"},"report_date":{"type":"string","format":"date","title":"Report Date"},"sb_opened":{"type":"integer","title":"Sb Opened","default":0},"sb_closed":{"type":"integer","title":"Sb Closed","default":0},"net_accounts":{"type":"integer","title":"Net Accounts","default":0},"pli_policies":{"type":"integer","title":"Pli Policies","default":0},"sum_assured":{"type":"number","title":"Sum Assured","default":0.0},"premium":{"type":"number","title":"Premium","default":0.0},"speed_post_document":{"type":"integer","title":"Speed Post Document","default":0},"speed_post_parcel":{"type":"integer","title":"Speed Post Parcel","default":0},"business_post":{"type":"integer","title":"Business Post","default":0},"logistics":{"type":"integer","title":"Logistics","default":0},"international_letter":{"type":"integer","title":"International Letter","default":0},"aadhaar_transactions":{"type":"integer","title":"Aadhaar Transactions","default":0},"aadhaar_amount":{"type":"number","title":"Aadhaar Amount","default":0.0},"id":{"type":"integer","title":"Id"}},"type":"object","required":["office_id","office_name","report_date","id"],"title":"DailyReportResponse"},"DailyReportSummary":{"properties":{"total_offices":{"type":"integer","title":"Total Offices"},"total_sb_opened":{"type":"integer","title":"Total Sb Opened"},"total_sb_closed":{"type":"integer","title":"Total Sb Closed"},"total_net_accounts":{"type":"integer","title":"Total Net Accounts"},"total_pli_policies":{"type":"integer","title":"Total Pli Policies"},"total_sum_assured":{"type":"number","title":"Total Sum Assured"},"total_premium":{"type":"number","title":"Total Premium"},"total_revenue":{"type":"number","title":"Total Revenue"},"report_date":{"type":"string","format":"date","title":"Report Date"}},"type":"object","required":["total_offices","total_sb_opened","total_sb_closed","total_net_accounts","total_pli_policies","total_sum_assured","total_premium","total_revenue","report_date"],"title":"DailyReportSummary"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LoginRequest":{"properties":{"employee_id":{"type":"string","title":"Employee Id"},"password":{"type":"string","title":"Password"}},"type":"object","required":["employee_id","password"],"title":"LoginRequest"},"OfficeBulkImport":{"properties":{"offices":{"items":{"$ref":"#/components/schemas/OfficeCreate"},"type":"array","title":"Offices"}},"type":"object","required":["offices"],"title":"OfficeBulkImport"},"OfficeCreate":{"properties":{"office_code":{"type":"string","maxLength":20,"minLength":2,"title":"Office Code","description":"Unique office code"},"office_name":{"type":"string","maxLength":255,"minLength":2,"title":"Office Name"},"office_type":{"$ref":"#/components/schemas/OfficeType"},"parent_office_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Parent Office Id"},"division":{"type":"string","title":"Division","default":"Nagpur City"},"region":{"type":"string","title":"Region","default":"Nagpur"},"circle":{"type":"string","title":"Circle","default":"Maharashtra"},"pincode":{"type":"string","maxLength":10,"minLength":6,"title":"Pincode"},"district":{"type":"string","title":"District"},"taluka":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Taluka"},"address":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Address"},"latitude":{"anyOf":[{"type":"number","maximum":90.0,"minimum":-90.0},{"type":"null"}],"title":"Latitude"},"longitude":{"anyOf":[{"type":"number","maximum":180.0,"minimum":-180.0},{"type":"null"}],"title":"Longitude"},"phone":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phone"},"email":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Email"},"beat_number":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Beat Number"},"jurisdiction_area":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Jurisdiction Area"},"status":{"allOf":[{"$ref":"#/components/schemas/OfficeStatus"}],"default":"ACTIVE"},"is_delivery_office":{"type":"boolean","title":"Is Delivery Office","default":true},"working_hours":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Working Hours"}},"type":"object","required":["office_code","office_name","office_type","pincode","district"],"title":"OfficeCreate"},"OfficeStatus":{"type":"string","enum":["ACTIVE","INACTIVE","TEMP_CLOSED","PERM_CLOSED"],"title":"OfficeStatus"},"OfficeType":{"type":"string","enum":["HEAD_OFFICE","SUB_OFFICE","BRANCH_OFFICE","ADMIN_OFFICE","OTHER"],"title":"OfficeType"},"OfficeUpdate":{"properties":{"office_name":{"anyOf":[{"type":"string","maxLength":255,"minLength":2},{"type":"null"}],"title":"Office Name"},"office_type":{"anyOf":[{"$ref":"#/components/schemas/OfficeType"},{"type":"null"}]},"parent_office_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Parent Office Id"},"division":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Division"},"region":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Region"},"circle":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Circle"},"pincode":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Pincode"},"district":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"District"},"taluka":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Taluka"},"address":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Address"},"latitude":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Latitude"},"longitude":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Longitude"},"phone":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phone"},"email":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Email"},"beat_number":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Beat Number"},"jurisdiction_area":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Jurisdiction Area"},"status":{"anyOf":[{"$ref":"#/components/schemas/OfficeStatus"},{"type":"null"}]},"is_delivery_office":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Delivery Office"},"working_hours":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Working Hours"}},"type":"object","title":"OfficeUpdate"},"RefreshTokenRequest":{"properties":{"refresh_token":{"type":"string","title":"Refresh Token","description":"Refresh token"}},"type":"object","required":["refresh_token"],"title":"RefreshTokenRequest"},"RegisterRequest":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"full_name":{"type":"string","maxLength":255,"minLength":2,"title":"Full Name"},"password":{"type":"string","minLength":8,"title":"Password"}},"type":"object","required":["email","full_name","password"],"title":"RegisterRequest"},"ReportRequest":{"properties":{"report_type":{"type":"string","pattern":"^(DAILY|MONTHLY|QUARTERLY|YEARLY|OFFICE_WISE|EMPLOYEE_WISE|SCHEME_WISE)$","title":"Report Type"},"filters":{"$ref":"#/components/schemas/AnalyticsFilter"},"format":{"type":"string","pattern":"^(JSON|PDF|EXCEL|CSV)$","title":"Format","default":"JSON"},"include_charts":{"type":"boolean","title":"Include Charts","default":true}},"type":"object","required":["report_type","filters"],"title":"ReportRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}